refactor(user): change name for helpers function(NO-JIRA)
